lossless-claw-6a6: extract engine payload externalization
